FIXME: PCI: make pci_bus_type const
Now that the driver core can properly handle constant struct bus_type,
move the pci_bus_type variable to be a constant structure as well,
placing it into read-only memory which can not be modified at runtime.

************************FIXME *********************************************
This really doesn't do the marking, drivers/iommu/iommu.c wants
to write to the iommu_ops structure of a number of bus types, including
the pci_bus_type, so that needs to be fixed up first before this
can be done properly.
************************FIXME *********************************************

Cc: Bjorn Helgaas <bhelgaas@google.com>
Cc: linux-pci@vger.kernel.org
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>