FIXME: gpu: host1x: make host1x_context_device_bus_type constant
Now that the driver core can properly handle constant struct bus_type,
move the host1x_context_device_bus_type variable to be a constant
structure as well, placing it into read-only memory which can not be
modified at runtime.

************************FIXME *********************************************
This really doesn't do the marking, drivers/iommu/iommu.c wants to write
to the iommu_ops structure of a number of bus types, including the
host1x_context_device_bus_type, so that needs to be fixed up first
before this can be done properly.
************************FIXME *********************************************

Cc: Thierry Reding <thierry.reding@gmail.com>
Cc: Mikko Perttunen <mperttunen@nvidia.com>
Cc: David Airlie <airlied@gmail.com>
Cc: Daniel Vetter <daniel@ffwll.ch>
Cc: dri-devel@lists.freedesktop.org
Cc: linux-tegra@vger.kernel.org
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>